Put test helpers in conftest.py
