+++ test_naucse/__init__.py
@@ -1,0 +1,6 @@
+# This is a package.
+
+# For a long time, pytest discouraged test packages, but no longer;  see:
+# https://docs.pytest.org/en/latest/goodpractices.html#tests-outside-application-code
+# This makes it easy for test_naucse.conftest to also include helpers that
+# aren't fixtures.
