+++ test_naucse/fixtures/expected-dumps/minimal-root.yml
@@ -3,7 +3,13 @@ api_version:
 - 0
 - 0
 root:
-    licenses: {}
+    licenses:
+        cc-by-sa-40:
+            title: Creative Commons Attribution-ShareAlike 4.0 International
+            url: https://creativecommons.org/licenses/by-sa/4.0/
+        cc0:
+            title: CC0 1.0 Universal Public Domain Dedication
+            url: https://creativecommons.org/publicdomain/zero/1.0/
     run_years: {}
     self_study_courses:
         courses/minimal:
